Replace diff mod with morph
