+++ crates/core/src/diff/diff.rs
@@ -1,509 +1,722 @@
-use std::collections::{BTreeMap, VecDeque};
+use std::cmp::Ordering;
+use std::collections::BTreeMap;
+use std::collections::BTreeSet;
+use std::collections::VecDeque;
+use std::fmt;
+use std::mem;
 
-use itertools::unfold;
 use smallvec::{smallvec, SmallVec};
+use std::ops::Deref;
 
 use crate::dom::*;
 
-use super::patch::Patch;
-use super::traversal::MoveTo;
-
-#[derive(Debug, Clone)]
-struct Cursor {
+use super::MoveTo;
+use super::Patch;
+#[derive(Clone)]
+struct Cursor<'a> {
+    doc: &'a Document,
     path: SmallVec<[u16; 6]>,
     node: NodeRef,
 }
-impl Cursor {
-    #[inline(always)]
-    fn node<'a>(&self, doc: &'a Document) -> &'a Node {
-        doc.get(self.node)
+
+impl fmt::Debug for Cursor<'_> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("Cursor")
+            .field("path", &self.path)
+            .field("node", &self.node)
+            .field("*node", &self.node())
+            .finish()
+    }
+}
+
+impl<'a> Cursor<'a> {
+    fn new(doc: &'a Document, node: NodeRef) -> Self {
+        Cursor {
+            doc,
+            path: smallvec![],
+            node,
+        }
     }
 
-    #[inline]
+    fn node(&self) -> &Node {
+        self.doc.get(self.node)
+    }
+
     fn depth(&self) -> usize {
         self.path.len()
     }
 
-    /// Moves the cursor up the tree to its parent, if it has one
-    fn up(&mut self, doc: &Document) {
-        if let Some(node) = doc.parent(self.node) {
+    // Create a cursor with the current node that iterates over descendant nodes
+    fn fork(&self) -> Cursor<'a> {
+        Cursor {
+            doc: self.doc,
+            path: smallvec![],
+            node: self.node,
+        }
+    }
+
+    // Create a cursor at `node` that iterates over descendant nodes
+    fn at(&self, node: NodeRef) -> Cursor<'a> {
+        Cursor::new(&self.doc, node)
+    }
+
+    fn children(&self) -> &'a [NodeRef] {
+        self.doc.children(self.node)
+    }
+
+    fn next_sibling(&self) -> Option<Cursor<'a>> {
+        let index = *self.path.last()? + 1;
+        let parent = self.doc.parent(self.node)?;
+        let node = *self.doc.children(parent).get(index as usize)?;
+        let mut path = self.path.clone();
+        *path.last_mut()? = index;
+
+        Some(Cursor {
+            doc: self.doc,
+            path,
+            node,
+        })
+    }
+
+    fn move_to_parent(&mut self) -> Option<()> {
+        if !self.path.is_empty() {
+            let parent = self.doc.parent(self.node)?;
+            self.node = parent;
             self.path.pop();
-            self.node = node;
+            Some(())
         } else {
-            panic!("expected parent");
+            None
         }
     }
 
-    /// Moves the cursor to the next sibling of the cursor's node, if possible
-    #[allow(unused)]
-    fn forward(&mut self, n: usize, doc: &Document) {
-        if let Some(node) = doc.parent(self.node) {
-            let index = self.path.last().copied().unwrap() as usize;
-            let pos = index + n;
-            let children = doc.children(node);
-            if let Some(node) = children.get(pos) {
-                self.path.pop();
-                self.path.push(pos as u16);
+    // Advance in depth-first order
+    fn advance(&mut self, skip_children: bool) -> Option<()> {
+        if !skip_children {
+            if let Some(node) = self.children().first() {
+                self.path.push(0);
                 self.node = *node;
-            } else {
-                panic!("expected child");
-            }
-        } else {
-            panic!("expected parent");
-        }
-    }
-}
-impl Eq for Cursor {}
-impl PartialEq for Cursor {
-    #[inline]
-    fn eq(&self, other: &Self) -> bool {
-        self.path.eq(&other.path)
-    }
-}
-impl PartialOrd for Cursor {
-    #[inline]
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-impl Ord for Cursor {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        use std::cmp::Ordering;
 
-        let mut anext = self.path.iter();
-        let mut bnext = other.path.iter();
+                return Some(());
+            }
+        }
+
+        let mut path = self.path.clone();
+        let mut parent = self.doc.parent(self.node)?;
 
         loop {
-            match (anext.next(), bnext.next()) {
-                (None, None) => break,
-                (None, _) => return Ordering::Less,
-                (_, None) => return Ordering::Greater,
-                (a, b) => match a.cmp(&b) {
-                    Ordering::Equal => continue,
-                    other => return other,
-                },
+            let index = *path.last()? + 1;
+
+            if let Some(node) = self.doc.children(parent).get(index as usize) {
+                *path.last_mut()? = index;
+
+                self.path = path;
+                self.node = *node;
+
+                return Some(());
+            } else {
+                if path.len() <= 1 {
+                    return None;
+                }
+
+                path.pop();
+
+                parent = self.doc.parent(parent)?;
+            }
+        }
+    }
+
+    fn advance_with_skip_list(
+        &mut self,
+        skip_children: bool,
+        skip_list: &BTreeSet<NodeRef>,
+    ) -> Option<()> {
+        let mut cursor = self.clone();
+
+        while cursor.advance(skip_children).is_some() {
+            if !skip_list.contains(&cursor.node) {
+                std::mem::swap(self, &mut cursor);
+                return Some(());
             }
         }
 
-        Ordering::Equal
+        None
+    }
+}
+
+impl<'a> Iterator for Cursor<'a> {
+    type Item = ();
+
+    fn next(&mut self) -> Option<()> {
+        self.advance(false)
+    }
+}
+
+impl<'a> From<&'a Document> for Cursor<'a> {
+    fn from(doc: &'a Document) -> Self {
+        Cursor::new(doc, doc.root())
+    }
+}
+
+impl<'a> Deref for Cursor<'a> {
+    type Target = Node;
+
+    fn deref(&self) -> &Self::Target {
+        self.node()
+    }
+}
+
+trait CompatibleWith: Deref<Target = Node> {
+    fn is_compatible_with<T>(&self, other: &T) -> bool
+    where
+        T: Deref<Target = Node>,
+    {
+        match (self.deref(), other.deref()) {
+            (Node::Element(from), Node::Element(to)) => {
+                to.name.eq(&from.name) && to.id().eq(&from.id())
+            }
+            (Node::Leaf(_), Node::Leaf(_)) => true,
+            (Node::Root, Node::Root) => true,
+            _ => false,
+        }
+    }
+}
+
+impl<T> CompatibleWith for T where T: Deref<Target = Node> {}
+
+#[derive(Debug)]
+enum Op<'a> {
+    Continue,
+    /// Detach node and keyed descendants then remove node
+    RemoveNode {
+        node: NodeRef,
+        /// A forked cursor of the node to be removed that iterates over descendant nodes
+        cursor: Cursor<'a>,
+        to: Cursor<'a>,
+        detach: bool,
+    },
+    /// Remove all `from` nodes
+    RemoveNodes {
+        from: Cursor<'a>,
+        to: Cursor<'a>,
+    },
+    /// Append `cursor` relative to the parent node last pushed to the stack by a patch operation
+    Append {
+        from: Cursor<'a>,
+        cursor: Cursor<'a>,
+    },
+    /// Append `to`
+    AppendNodes {
+        from: Cursor<'a>,
+        to: Cursor<'a>,
+    },
+    /// Append sibling nodes relative to the parent node on the stack
+    AppendSiblings {
+        from: Cursor<'a>,
+        cursor: Cursor<'a>,
+    },
+    /// Inserts `node` before the current node and appends descendant nodes
+    InsertBefore {
+        from: Cursor<'a>,
+        cursor: Cursor<'a>,
+    },
+    /// Detach node if not already detached
+    MaybeDetach {
+        node: NodeRef,
+    },
+    Morph(Cursor<'a>, Cursor<'a>),
+    Patch(Patch),
+}
+
+impl Default for Op<'_> {
+    fn default() -> Self {
+        Op::Continue
+    }
+}
+
+impl<'a, T> From<(T, T)> for Op<'a>
+where
+    T: Into<Cursor<'a>>,
+{
+    fn from((from, to): (T, T)) -> Self {
+        Op::Morph(from.into(), to.into())
+    }
+}
+
+enum Advance {
+    BothCursors,
+    To,
+    From,
+}
+
+struct Morph<'a> {
+    stack: SmallVec<[Op<'a>; 16]>,
+    queue: SmallVec<[Op<'a>; 8]>,
+    detached: BTreeSet<NodeRef>,
+}
+
+impl<'a> Morph<'a> {
+    fn new(from: &'a Document, to: &'a Document) -> Self {
+        (from, to).into()
+    }
+
+    // enqueue patches to morph attributes
+    fn morph_current_attr(&mut self) {
+        if let Some(Op::Morph(from, to)) = self.stack.last_mut() {
+            let node = from.node;
+
+            if let (Node::Element(old), Node::Element(new)) = (from.node(), to.node()) {
+                let mut current = BTreeMap::from_iter(
+                    old.attributes()
+                        .into_iter()
+                        .map(|attr| (&attr.name, &attr.value)),
+                );
+
+                self.queue
+                    .extend(new.attributes().into_iter().filter_map(|attr| {
+                        match current.remove(&attr.name) {
+                            Some(value) if value.ne(&attr.value) => {
+                                Some(Op::Patch(Patch::UpdateAttribute {
+                                    node,
+                                    name: attr.name.to_owned(),
+                                    value: attr.value.to_owned(),
+                                }))
+                            }
+                            Some(_) => None,
+                            None => Some(Op::Patch(Patch::AddAttributeTo {
+                                node,
+                                name: attr.name.to_owned(),
+                                value: attr.value.to_owned(),
+                            })),
+                        }
+                    }));
+
+                while let Some(patch) =
+                    current
+                        .pop_first()
+                        .map(|(name, _)| Patch::RemoveAttributeByName {
+                            node,
+                            name: name.to_owned(),
+                        })
+                {
+                    self.queue.push(Op::Patch(patch));
+                }
+            }
+        }
+    }
+
+    fn advance(&mut self, advance: Advance, skip_children: bool) {
+        let op = self.stack.last_mut().unwrap();
+
+        if let Op::Morph(from, to) = op {
+            let next = match advance {
+                Advance::BothCursors => (
+                    from.advance_with_skip_list(skip_children, &self.detached),
+                    to.advance(skip_children),
+                ),
+                Advance::To => (Some(()), to.advance(skip_children)),
+                Advance::From => (
+                    from.advance_with_skip_list(skip_children, &self.detached),
+                    Some(()),
+                ),
+            };
+
+            match next {
+                (Some(_), Some(_)) => {}
+                (Some(_), None) => {
+                    *op = Op::RemoveNodes {
+                        from: from.to_owned(),
+                        to: to.to_owned(),
+                    };
+                }
+                (None, Some(_)) => {
+                    *op = Op::AppendNodes {
+                        from: from.to_owned(),
+                        to: to.to_owned(),
+                    };
+                }
+                (None, None) => {
+                    *op = Op::Continue;
+                }
+            }
+        }
+    }
+}
+
+impl<'a, T> From<T> for Morph<'a>
+where
+    T: Into<Op<'a>>,
+{
+    fn from(op: T) -> Self {
+        Morph {
+            stack: smallvec![op.into()],
+            queue: smallvec![],
+            detached: BTreeSet::new(),
+        }
+    }
+}
+
+impl<'a> Iterator for Morph<'a> {
+    type Item = Patch;
+
+    fn next(&mut self) -> Option<Patch> {
+        loop {
+            while let Some(Op::Continue) = self.stack.last() {
+                self.stack.pop();
+            }
+
+            self.stack.extend(self.queue.drain(..).rev());
+
+            let op = self.stack.last_mut()?;
+
+            match op {
+                Op::Continue => {
+                    unreachable!("Op::Continue should be popped off the stack prior")
+                }
+                Op::RemoveNode {
+                    ref node,
+                    cursor,
+                    ref to,
+                    detach,
+                } => {
+                    let node = *node;
+
+                    if *detach {
+                        *detach = false;
+
+                        if !cursor.children().is_empty() {
+                            self.queue.push(Op::Patch(Patch::Detach { node }));
+                            continue;
+                        }
+                    }
+
+                    if cursor.next().is_some() {
+                        if let Node::Element(el) = cursor.node() {
+                            if let Some(id) = el.id() {
+                                if to.doc.get_by_id(id).is_some() {
+                                    // Only detach if not previously moved
+                                    if self.detached.insert(cursor.node) {
+                                        self.queue
+                                            .push(Op::Patch(Patch::Detach { node: cursor.node }));
+                                        continue;
+                                    }
+                                }
+                            }
+                        }
+                    }
+
+                    *op = Op::Patch(Patch::Remove { node });
+                }
+                Op::RemoveNodes { from, to } => {
+                    if !self.detached.contains(&from.node) {
+                        self.queue.push(Op::RemoveNode {
+                            node: from.node,
+                            cursor: from.fork(),
+                            to: to.fork(),
+                            detach: true,
+                        });
+                    }
+
+                    if from.advance(true).is_none() {
+                        *op = Op::Continue;
+                    }
+                }
+                Op::Append { ref from, cursor } => {
+                    if let Some(id) = cursor.id() {
+                        if let Some(node) = from.doc.get_by_id(id) {
+                            self.queue.extend([
+                                Op::MaybeDetach { node },
+                                // Parent will already be on the stack so only need to push child
+                                Op::Patch(Patch::Push(node)),
+                                // Attach will pop node off the stack and append to parent
+                                Op::Patch(Patch::Attach),
+                                Op::Morph(from.at(node), cursor.fork()),
+                            ]);
+
+                            if cursor.advance(true).is_none() {
+                                *op = Op::Continue;
+                            }
+
+                            continue;
+                        }
+                    }
+
+                    let node = cursor.node().to_owned();
+                    let depth = cursor.depth();
+
+                    if cursor.next().is_some() {
+                        match cursor.depth().cmp(&depth) {
+                            Ordering::Less => {
+                                unreachable!(
+                                    "cursor should always be forked so that it will never move up"
+                                );
+                            }
+                            // Next node is also a sibling, so leave parent unchanged and append current
+                            Ordering::Equal => {
+                                self.queue.push(Op::Patch(Patch::Append { node }));
+                                continue;
+                            }
+                            // Next node is a child node, so append and make node the new parent
+                            Ordering::Greater => {
+                                self.queue.extend([
+                                    Op::Patch(Patch::CreateAndMoveTo { node }),
+                                    // Insertion point resets to parent
+                                    Op::Patch(Patch::Attach),
+                                    // Move to newly created node relative to parent
+                                    Op::Patch(Patch::Move(MoveTo::ReverseChild(0))),
+                                    // Set created node as parent for next append
+                                    Op::Patch(Patch::PushCurrent),
+                                    Op::Append {
+                                        from: from.clone(),
+                                        cursor: cursor.fork(),
+                                    },
+                                    Op::Patch(Patch::Pop),
+                                ]);
+
+                                if cursor.advance(true).is_none() {
+                                    *op = Op::Continue;
+                                }
+                            }
+                        }
+                    } else {
+                        // self.queue.push(Op::Patch(Patch::Append { node }));
+
+                        self.queue.extend([
+                            Op::Patch(Patch::CreateAndMoveTo { node }),
+                            // Insertion point resets to parent
+                            Op::Patch(Patch::Attach),
+                        ]);
+
+                        *op = Op::Continue;
+                    }
+                }
+                Op::AppendNodes { from, to } => {
+                    // Move cursor to parent for append
+                    while from.depth().ge(&to.depth()) {
+                        from.move_to_parent();
+                    }
+
+                    self.queue.extend([
+                        Op::Patch(Patch::Push(from.node)),
+                        Op::AppendSiblings {
+                            from: from.clone(),
+                            cursor: to.clone(),
+                        },
+                        Op::Patch(Patch::Pop),
+                    ]);
+
+                    while to.move_to_parent().is_some() {
+                        if to.advance(true).is_some() {
+                            continue;
+                        }
+                    }
+
+                    *op = Op::Continue;
+                }
+                Op::AppendSiblings {
+                    ref from,
+                    cursor: to,
+                } => {
+                    self.queue.push(Op::Append {
+                        from: from.clone(),
+                        cursor: to.fork(),
+                    });
+
+                    if let Some(next) = to.next_sibling() {
+                        *to = next;
+                    } else {
+                        *op = Op::Continue;
+                    }
+                }
+                Op::InsertBefore { ref from, cursor } => {
+                    let node = cursor.node().to_owned();
+
+                    if cursor.next().is_some() {
+                        self.queue.extend([
+                            Op::Patch(Patch::CreateAndMoveTo { node }),
+                            Op::Patch(Patch::PrependBefore { before: from.node }),
+                            // Set newly inserted node as append parent
+                            Op::Patch(Patch::PushCurrent),
+                            Op::Append {
+                                from: from.clone(),
+                                cursor: cursor.fork(),
+                            },
+                            Op::Patch(Patch::Pop),
+                        ]);
+
+                        *op = Op::Continue;
+                    } else {
+                        *op = Op::Patch(Patch::InsertBefore {
+                            before: from.node,
+                            node,
+                        });
+                    }
+                }
+                Op::MaybeDetach { ref node } => {
+                    // If a node was previously moved, it will have already been set as detached and can be ignored
+                    if self.detached.insert(*node) {
+                        *op = Op::Patch(Patch::Detach { node: *node });
+                    } else {
+                        *op = Op::Continue;
+                    }
+                }
+                Op::Morph(from, to) => {
+                    // paths are relative; a forked cursor without siblings will have a depth of 0
+                    match to.depth().cmp(&from.depth()) {
+                        // node added as child
+                        Ordering::Greater => {
+                            self.queue.extend([
+                                Op::Patch(Patch::Push(from.node)),
+                                Op::Append {
+                                    from: from.clone(),
+                                    cursor: to.fork(),
+                                },
+                                Op::Patch(Patch::Pop),
+                            ]);
+
+                            self.advance(Advance::To, true);
+                            continue;
+                        }
+                        Ordering::Equal => {}
+                        // existing node deleted
+                        Ordering::Less => {
+                            self.queue.push(Op::RemoveNode {
+                                node: from.node,
+                                cursor: from.fork(),
+                                to: to.fork(),
+                                detach: true,
+                            });
+
+                            self.advance(Advance::From, true);
+                            continue;
+                        }
+                    }
+
+                    match (from.node(), to.node()) {
+                        (Node::Root, Node::Root) | (Node::Root, _) | (_, Node::Root) => {
+                            self.advance(Advance::BothCursors, false);
+                        }
+                        (Node::Leaf(old_content), Node::Leaf(content)) => {
+                            if old_content.ne(content) {
+                                self.queue.push(Op::Patch(Patch::Replace {
+                                    node: from.node,
+                                    replacement: Node::Leaf(content.to_owned()),
+                                }));
+                            }
+
+                            self.advance(Advance::BothCursors, false);
+                        }
+                        (Node::Leaf(_), Node::Element(_)) => {
+                            self.queue
+                                .push(Op::Patch(Patch::Remove { node: from.node }));
+
+                            self.advance(Advance::From, true);
+                        }
+                        (Node::Element(_), Node::Leaf(content)) => {
+                            self.queue.push(Op::Patch(Patch::InsertBefore {
+                                before: from.node,
+                                node: Node::Leaf(content.to_owned()),
+                            }));
+
+                            self.advance(Advance::To, true);
+                        }
+                        (Node::Element(from_el), Node::Element(to_el)) => {
+                            // nodes are compatible; morph attribute changes and continue
+                            if to_el.name.eq(&from_el.name) && to_el.id().eq(&from_el.id()) {
+                                self.morph_current_attr();
+
+                                self.advance(Advance::BothCursors, false);
+                                continue;
+                            }
+
+                            // Keyed node shouldn't be here; detach/remove and continue
+                            if let Some(id) = from.id() {
+                                if to.doc.get_by_id(id).is_some() {
+                                    self.queue.push(Op::MaybeDetach { node: from.node });
+                                } else {
+                                    self.queue.push(Op::RemoveNode {
+                                        node: from.node,
+                                        cursor: from.fork(),
+                                        to: to.fork(),
+                                        detach: true,
+                                    });
+                                }
+
+                                self.advance(Advance::From, true);
+                                continue;
+                            }
+
+                            // If keyed el should be here, relocated or insert instead of transforming el
+                            if let Some(id) = to.id() {
+                                if let Some(node) = from.doc.get_by_id(id) {
+                                    self.queue.extend([
+                                        Op::Patch(Patch::Push(node)),
+                                        Op::MaybeDetach { node },
+                                        Op::Patch(Patch::PrependBefore { before: from.node }),
+                                        Op::Morph(from.at(node), to.fork()),
+                                    ]);
+
+                                    continue;
+                                } else {
+                                    self.queue.push(Op::InsertBefore {
+                                        from: from.clone(),
+                                        cursor: to.fork(),
+                                    });
+
+                                    self.advance(Advance::To, true);
+                                    continue;
+                                }
+                            }
+
+                            // If the next existing el can be morphed into the target el, delete current instead of replacing
+                            if let Some(from_next) = from.next_sibling() {
+                                if from_next.is_compatible_with(to) {
+                                    self.queue.push(Op::RemoveNode {
+                                        node: from.node,
+                                        cursor: from.fork(),
+                                        to: to.fork(),
+                                        detach: true,
+                                    });
+
+                                    self.advance(Advance::From, true);
+                                    continue;
+                                }
+                            }
+
+                            // If the next node being morphed into is compatible, insert target node before current
+                            if let Some(to_next) = to.next_sibling() {
+                                if to_next.is_compatible_with(from) {
+                                    self.queue.push(Op::InsertBefore {
+                                        from: from.clone(),
+                                        cursor: to.fork(),
+                                    });
+
+                                    self.advance(Advance::To, true);
+                                    continue;
+                                }
+                            }
+
+                            // TODO: as an optimization, use peek to add node caching
+                            self.queue.push(Op::Patch(Patch::Replace {
+                                node: from.node,
+                                replacement: Node::Element(to_el.to_owned()),
+                            }));
+
+                            self.advance(Advance::BothCursors, false);
+                        }
+                    }
+                }
+                Op::Patch(_) => {
+                    if let Op::Patch(patch) = mem::replace(op, Op::Continue) {
+                        return Some(patch);
+                    } else {
+                        unreachable!();
+                    }
+                }
+            }
+        }
     }
 }
 
 pub fn diff(old_document: &Document, new_document: &Document) -> VecDeque<Patch> {
-    let mut patches = VecDeque::new();
-    let mut old_next = VecDeque::from([Cursor {
-        path: smallvec![],
-        node: old_document.root(),
-    }]);
-    let mut new_next = VecDeque::from([Cursor {
-        path: smallvec![],
-        node: new_document.root(),
-    }]);
-    loop {
-        match (old_next.pop_front(), new_next.pop_front()) {
-            // We're at the same position in both trees, so examine the details of the node under the cursor
-            (Some(old_cursor), Some(new_cursor)) if old_cursor == new_cursor => {
-                match (old_cursor.node(old_document), new_cursor.node(new_document)) {
-                    // This was the root node, so move the cursor down a level and start walking the children
-                    (Node::Root, Node::Root) => {
-                        let old_children = old_document.children(old_cursor.node);
-                        let new_children = new_document.children(new_cursor.node);
-                        old_next.extend(old_children.iter().copied().enumerate().map(
-                            |(i, node)| {
-                                let mut path = old_cursor.path.clone();
-                                path.push(i as u16);
-                                Cursor { path, node }
-                            },
-                        ));
-                        new_next.extend(new_children.iter().copied().enumerate().map(
-                            |(i, node)| {
-                                let mut path = new_cursor.path.clone();
-                                path.push(i as u16);
-                                Cursor { path, node }
-                            },
-                        ));
-                    }
-                    // Both nodes are leaf nodes, compare their content for equality
-                    (Node::Leaf(ref ol), Node::Leaf(ref nl)) => {
-                        // If both nodes are identical, keep moving
-                        // Otherwise, emit a patch to update the text of this node
-                        if ol == nl {
-                            continue;
-                        } else {
-                            patches.push_back(Patch::Replace {
-                                node: old_cursor.node,
-                                replacement: Node::Leaf(nl.clone()),
-                            });
-                        }
-                    }
-                    // Both nodes are elements, compare their tag and attributes for equality, then append their children to the stack
-                    (Node::Element(ref old), Node::Element(ref new)) => {
-                        // If the names are different, replace the old element with the new
-                        if old.name != new.name {
-                            patches.push_back(Patch::Replace {
-                                node: old_cursor.node,
-                                replacement: Node::Element(new.clone()),
-                            });
-                        } else {
-                            // Check for changes to the attributes
-                            patches.extend(diff_attributes(old_cursor.node, old, new));
-                        }
-                        // Add the children of both nodes to the worklist
-                        let old_children = old_document.children(old_cursor.node);
-                        let new_children = new_document.children(new_cursor.node);
-                        old_next.extend(old_children.iter().copied().enumerate().map(
-                            |(i, node)| {
-                                let mut path = old_cursor.path.clone();
-                                path.push(i as u16);
-                                Cursor { path, node }
-                            },
-                        ));
-                        new_next.extend(new_children.iter().copied().enumerate().map(
-                            |(i, node)| {
-                                let mut path = new_cursor.path.clone();
-                                path.push(i as u16);
-                                Cursor { path, node }
-                            },
-                        ));
-                    }
-                    // The old node was an element and the new node is a leaf; determine if this is a simple swap, addition, or removal
-                    // by looking forward in the stack to future cursors which are at the same depth.
-                    (Node::Element(_), Node::Leaf(ref new)) => {
-                        patches.push_back(Patch::Replace {
-                            node: old_cursor.node,
-                            replacement: Node::Leaf(new.clone()),
-                        });
-                        let old_children = old_document.children(old_cursor.node);
-                        old_next.extend(old_children.iter().copied().enumerate().map(
-                            |(i, node)| {
-                                let mut path = old_cursor.path.clone();
-                                path.push(i as u16);
-                                Cursor { path, node }
-                            },
-                        ));
-                    }
-                    // The old node was a leaf and the new node is an element; determine if this is a simple swap, addition, or removal
-                    // by looking forward in the stack to future cursors which are at the same depth.
-                    (Node::Leaf(_), Node::Element(ref new)) => {
-                        patches.push_back(Patch::Replace {
-                            node: old_cursor.node,
-                            replacement: Node::Element(new.clone()),
-                        });
-                        let new_children = new_document.children(new_cursor.node);
-                        new_next.extend(new_children.iter().copied().enumerate().map(
-                            |(i, node)| {
-                                let mut path = new_cursor.path.clone();
-                                path.push(i as u16);
-                                Cursor { path, node }
-                            },
-                        ));
-                    }
-                    _ => unreachable!(),
-                }
-            }
-            // This occurs in the following scenario (where `^` indicates the cursor location in each tree):
-            //
-            //       old     new                  old worklist  new worklist
-            //        |       |   <- depth 0      a             a
-            //       / \     / \                  b             b
-            //      a   b   a   b <- depth 1      c             c
-            //     / \  |   |   |                 d <           e <
-            //    c  d  e   c   e <- depth 2      e
-            //       ^          ^
-            //
-            // In this scenario, the new tree has had `d` removed from it, and as nodes are added to the worklist
-            // in breadth-first order, this means that when we move to `d` in the old tree, we'll move to `e` in the new.
-            //
-            // When this occurs, it means that the old parent node (e.g. `a` in the example above) has extra nodes which need
-            // to be removed. To bring the old cursor up to par with the new cursor, we remove all nodes in the old tree until
-            // we catch up to it
-            (Some(old_cursor), Some(new_cursor))
-                if old_cursor < new_cursor && old_cursor.depth() == new_cursor.depth() =>
-            {
-                patches.push_back(Patch::Remove {
-                    node: old_cursor.node,
-                });
-                while let Some(old_cursor) = old_next.front() {
-                    if old_cursor >= &new_cursor {
-                        // We've caught up to the new cursor
-                        break;
-                    }
-                    let old_cursor = old_next.pop_front().unwrap();
-                    patches.push_back(Patch::Remove {
-                        node: old_cursor.node,
-                    });
-                }
-                // We need to revisit the new_cursor now that we've adjusted the old cursor
-                new_next.push_front(new_cursor);
-            }
-            // This occurs in the following scenario (where `^` indicates the cursor location in each tree):
-            //
-            //       old      new                  old worklist  new worklist
-            //        |        |   <- depth 0      a             a
-            //       / \      / \                  b             b
-            //      a   b    a   b <- depth 1      c             c
-            //      |   |   / \  |                 d <           e <
-            //      c   d  c   e d <- depth 2                    d
-            //          ^      ^
-            //
-            // Similar to the previous scenario, this one occurs when the new tree has had `e` added to `a` as a new
-            // child element. This means that when we move to `d` in the old tree, we'll move to `e` in the new.
-            //
-            // As implied, this means that the old parent node (e.g. `a` in the example above) has nodes which need
-            // to be added to it. To bring the new cursor up to par with the old cursor, we add every node in the new tree
-            // until we've caught up to it
-            (Some(old_cursor), Some(new_cursor))
-                if old_cursor > new_cursor && old_cursor.depth() == new_cursor.depth() =>
-            {
-                patches.push_back(Patch::Move(MoveTo::Node(old_document.root())));
-                // Traverse the old tree based on the new_cursor path until we get to its immediate parent
-                {
-                    let (_, parent_path) = new_cursor.path.split_last().unwrap();
-                    for index in parent_path.iter().copied() {
-                        patches.push_back(Patch::Move(MoveTo::Child(index as u32)));
-                    }
-                }
-                patches.push_back(Patch::PushCurrent);
-                let mut subtree = recursively_append(new_cursor.node, new_document);
-                patches.append(&mut subtree);
-                old_next.push_front(old_cursor);
-            }
-            // This occurs in the following scenario (where `^` indicates the cursor location in each tree):
-            //
-            //       old      new                  old worklist  new worklist
-            //        |        |     <- depth 0      a             a
-            //       / \      / \                    b             b
-            //      a   b    a   b   <- depth 1      c             c
-            //     /        /   / \                  d <           f <
-            //    c        c   f   g <- depth 2      e             g
-            //    |        |   ^                                   d
-            //    d<       d   j                                   j
-            //    |        |                                       e
-            //    e        e
-            //
-            //
-            // In this scenario, a new child element was added in the new tree which will be visited first before we can catch up
-            // with the old cursor. To catch up, we simply add all nodes in the new tree until we reach the old cursor.
-            // This is like the previous scenario, but inverted. The new tree has been modified in such a way that when we move from
-            // `c` in both trees, we'll end up at a greater depth in the old tree.
-            //
-            // In order to proceed, we must add all nodes from the new tree with the same depth and parent as the initial new cursor,
-            // until we either reach the end of the new tree, or we catch up to the old cursor
-            (Some(old_cursor), Some(new_cursor))
-                if new_cursor > old_cursor && old_cursor.depth() > new_cursor.depth() =>
-            {
-                // Create a cursor to traverse back up the tree to the last common parent
-                let mut tmp_cursor = old_cursor.clone();
-                let hops = old_cursor.depth() - new_cursor.depth();
-                for _ in 0..hops {
-                    tmp_cursor.up(old_document);
-                }
-                debug_assert_eq!(tmp_cursor.depth(), new_cursor.depth());
-                // Insert the missing node after the current node in the old tree, as it must be an immediate sibling
-                let mut subtree =
-                    recursively_append_after(tmp_cursor.node, new_cursor.node, new_document);
-                patches.append(&mut subtree);
-                // Retry with the next item in the worklist
-                old_next.push_front(old_cursor);
-            }
-            // This is the exact inverse of the scenario above, where rather than a new child being introduced, a previously
-            // existing node has been removed, resulting in the old cursor moving ahead in the tree compared to the new cursor.
-            //
-            //       old        new                  old worklist  new worklist
-            //        |          |     <- depth 0      a             a
-            //       / \        / \                    b             b
-            //      a   b      a   b   <- depth 1      c             c
-            //     /   / \    /                        f <           d <
-            //    c   f   g  c         <- depth 2      g             e
-            //    |   ^      |                         d
-            //    d   j      d<                        j
-            //    |          |                         e
-            //    e          e
-            //
-            //           //
-            (Some(old_cursor), Some(new_cursor))
-                if old_cursor > new_cursor && new_cursor.depth() > old_cursor.depth() =>
-            {
-                patches.push_back(Patch::Remove {
-                    node: old_cursor.node,
-                });
-                // Retry with the next item in the worklist
-                new_next.push_front(new_cursor);
-            }
-            (Some(old_cursor), Some(new_cursor)) => {
-                panic!(
-                    "unexpected edge case in diff calculation: {:#?} vs {:#?}",
-                    old_cursor, new_cursor
-                );
-            }
-            // We've reached the end of the worklist for the new tree, but not the old; this means
-            // that all remaining nodes in the old tree were removed, since we can't have visited them yet,
-            // so issue removals for all remaining old tree nodes
-            (Some(old_cursor), None) => {
-                patches.push_back(Patch::Remove {
-                    node: old_cursor.node,
-                });
-                while let Some(old_cursor) = old_next.pop_front() {
-                    patches.push_back(Patch::Remove {
-                        node: old_cursor.node,
-                    });
-                }
-            }
-            // We've reached the end of the worklist for the old tree, which means that all remaining nodes in the
-            // new tree were added.
-            //
-            //      old          new         old worklist  new worklist
-            //       |            |            a             a
-            //      / \          / \           b             b
-            //     a   b        a   b          *<            c<
-            //                  |   |                        d
-            //                  c   d
-            //
-            // As shown above, additions must necessarily be rooted under nodes of the tree that already exist, but
-            // they might be under different parents. We use the cursor path to traverse to the appropriate node of
-            // the tree in which to append the new children.
-            (None, Some(new_cursor)) => {
-                patches.push_back(Patch::Move(MoveTo::Node(old_document.root())));
-                // Traverse the old tree based on the new_cursor path until we get to its immediate parent
-                // Skip the last index in the new_cursor path, because that's the index of the
-                // new_cursor node (c in the diagram above), which does not yet exist in the tree.
-                for index in new_cursor.path[..new_cursor.path.len() - 1].iter().copied() {
-                    patches.push_back(Patch::Move(MoveTo::Child(index as u32)));
-                }
-                patches.push_back(Patch::PushCurrent);
-                let mut subtree = recursively_append(new_cursor.node, new_document);
-                patches.append(&mut subtree);
-            }
-            // We've reached the end of the worklist, at the same time, we're done
-            (None, None) => break,
-        }
-    }
-
-    patches
-}
-
-pub fn diff_attributes<'a>(
-    node: NodeRef,
-    old: &'a Element,
-    new: &'a Element,
-) -> impl Iterator<Item = Patch> + 'a {
-    let current: BTreeMap<&'a AttributeName, &'a AttributeValue> = BTreeMap::from_iter(
-        old.attributes()
-            .into_iter()
-            .map(|attr| (&attr.name, &attr.value)),
-    );
-
-    unfold(
-        (current, new.attributes().into_iter()),
-        move |(current, new)| {
-            while let Some(attr) = new.next() {
-                match current.remove(&attr.name) {
-                    Some(value) if value.ne(&attr.value) => {
-                        return Some(Patch::UpdateAttribute {
-                            node,
-                            name: attr.name.to_owned(),
-                            value: attr.value.to_owned(),
-                        });
-                    }
-                    Some(_) => {}
-                    None => {
-                        return Some(Patch::AddAttributeTo {
-                            node,
-                            name: attr.name.to_owned(),
-                            value: attr.value.to_owned(),
-                        });
-                    }
-                }
-            }
-
-            current
-                .pop_first()
-                .map(|(name, _)| Patch::RemoveAttributeByName {
-                    node,
-                    name: name.to_owned(),
-                })
-        },
-    )
-}
-
-fn recursively_append_after(
-    after: NodeRef,
-    src: NodeRef,
-    src_document: &Document,
-) -> VecDeque<Patch> {
-    use petgraph::visit::{depth_first_search, DfsEvent};
-
-    let mut patches = VecDeque::new();
-
-    depth_first_search(src_document, Some(src), |event| {
-        match event {
-            DfsEvent::Discover(node, _) => {
-                match src_document.get(node) {
-                    Node::Leaf(ref s) => patches.push_back(Patch::CreateAndMoveTo {
-                        node: Node::Leaf(s.clone()),
-                    }),
-                    Node::Element(ref elem) => {
-                        patches.push_back(Patch::CreateAndMoveTo {
-                            node: Node::Element(elem.clone()),
-                        });
-                    }
-                    // Ignore the root
-                    Node::Root => (),
-                };
-            }
-            DfsEvent::TreeEdge(_from, _to) => {
-                // Moving down the tree one level, ignored because Discover takes care of traversing for us
-                ()
-            }
-            DfsEvent::Finish(node, _) if node == src => {
-                // All children of `node` have been visited, and we're returning up the tree
-                patches.push_back(Patch::AppendAfter { after });
-            }
-            DfsEvent::Finish(_node, _) => {
-                // All children of `node` have been visited, and we're returning up the tree
-                patches.push_back(Patch::Attach);
-            }
-            // This is a tree, these types of edges aren't allowed
-            DfsEvent::BackEdge(_, _) | DfsEvent::CrossForwardEdge(_, _) => unreachable!(),
-        }
-    });
-
-    patches
-}
-
-fn recursively_append(src: NodeRef, src_document: &Document) -> VecDeque<Patch> {
-    use petgraph::visit::{depth_first_search, DfsEvent};
-
-    let mut patches = VecDeque::new();
-
-    depth_first_search(src_document, Some(src), |event| {
-        match event {
-            DfsEvent::Discover(node, _) => {
-                match src_document.get(node) {
-                    Node::Leaf(ref s) => patches.push_back(Patch::CreateAndMoveTo {
-                        node: Node::Leaf(s.clone()),
-                    }),
-                    Node::Element(ref elem) => {
-                        patches.push_back(Patch::CreateAndMoveTo {
-                            node: Node::Element(elem.clone()),
-                        });
-                    }
-                    // Ignore the root
-                    Node::Root => (),
-                };
-            }
-            DfsEvent::TreeEdge(_from, _to) => {
-                // Moving down the tree one level, ignored because Discover takes care of traversing for us
-                ()
-            }
-            DfsEvent::Finish(_node, _) => {
-                // All children of `node` have been visited, and we're returning up the tree
-                patches.push_back(Patch::Attach);
-            }
-            // This is a tree, these types of edges aren't allowed
-            DfsEvent::BackEdge(_, _) | DfsEvent::CrossForwardEdge(_, _) => unreachable!(),
-        }
-    });
-
-    patches
+    VecDeque::from_iter(Morph::new(old_document, new_document))
 }
